+++ yue-library-data-redis/src/main/java/ai/yue/library/data/redis/dto/WxUserDTO.java
@@ -1,0 +1,28 @@
+package ai.yue.library.data.redis.dto;
+
+import com.alibaba.fastjson.JSONArray;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * @author  孙金川
+ * @version 创建时间：2018年9月11日
+ */
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class WxUserDTO {
+	
+	String openid;
+	String nickname;
+	Character sex;
+	String country;
+	String province;
+	String city;
+	String headimgurl;
+	JSONArray privilege;
+	String unionid;
+	
+}